% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nmadt.R
\name{print.nmadt}
\alias{print.nmadt}
\title{Print method for `nmadt` objects}
\usage{
\method{print}{nmadt}(x, ...)
}
\arguments{
\item{x}{An object of class `nmadt`}

\item{...}{Not used. Included for S3 method compatibility.}
}
\description{
Print method for `nmadt` objects
}
\examples{
\donttest{
data(dat.kang)
set.seed(9)
kang.out <- nmadt.hierarchical(nstu=12, K=2, data=dat.kang, 
            testname=c("D-dimer","Ultrasonography"))
print(kang.out) 
}

}
